Moved tests to new folder and updated it
